pyproject.toml
[tool.poetry]
name = "smitproject"
version = "0.1.0"
description = ""
authors = ["arturkalimullin <noti1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
tortoise-orm = "^0.19.3"
pydantic = "1.10.11"
asyncpg = "^0.28.0"
orjson = "^3.9.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
flake8-blind-except = "^0.2.1"
flake8-bugbear = "^23.7.10"
flake8-builtins = "^2.1.0"
flake8-class-attributes-order = "^0.1.3"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.14.0"
flake8-debugger = "^4.1.2"
flake8-functions = "^0.0.8"
flake8-isort = "^6.0.0"
flake8-mutable = "^1.2.0"
flake8-print = "^5.0.0"
flake8-pytest = "^1.4"
flake8-pytest-style = "^1.7.2"
flake8-quotes = "^3.3.2"
flake8-string-format = "^0.3.0"
flake8-variables-names = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
